+++ src/components/SpellSelection/index.tsx
@@ -1,5 +1,5 @@
 import CopyToClipboard from "react-copy-to-clipboard";
-import { DragSpell } from "components/Spell";
+import { DragSpell } from "components/Spell/DragSpell";
 import { Spells } from "lib";
 import { Spell as SpellType, SpellCategory } from "lib/types";
 import { Button } from "components/Button";
@@ -9,8 +9,6 @@ import { isSerializedSimulationState } from "context/simulations.selectors";
 import clipboard from "clipboardy";
 import lzbase62 from "lzbase62";
 import { importSimulation } from "context/simulations.actions";
-
-const RAMP_SEQUENCE = `ramp-uDritjZdDkRkVrisS2MtYVRcZeX2LsIEritURdRXxWDxLDtRSjfiSxLFtSlWWj2KtLN2Ltk25xfGtgcRuBtViyDEtjgVccgfnxOEsKI2B2ctYRjyeEsRRxMD3TxMDtiuB2OsLNxODtTiZkxLHtm2WyXDsMxwE2i4RsaynGySOy1EyQHxLEyPJxNEyOGxKEyNG2KuFFxWDtffcU3AtexRG2Q2ktVd0EHti5ltgDygExQEsMPNLOEJPxDY28`;
 
 function set<T>(s: string, o: { [index: string]: T[] }, i: T) {
   if (o[s]) {
@@ -67,7 +65,7 @@ export const SpellSelection = React.memo(function () {
           <Button outline icon="DownloadIcon" onClick={importString}>
             Import
           </Button>
-          <CopyToClipboard text={RAMP_SEQUENCE}>
+          <CopyToClipboard text={"No ramp available for Dragonflight right now 😥"}>
             <Button icon="ClipboardCopyIcon">Copy Ramp Sequence</Button>
           </CopyToClipboard>
         </div>
@@ -78,7 +76,7 @@ export const SpellSelection = React.memo(function () {
             <h5 className="text-md text-gray-600 font-semibold">{key}</h5>
             <div className="bg-gradient-to-b from-gray-100 to-gray-200 p-2 rounded drop-shadow-sm">
               {spells.map((spell) => (
-                <DragSpell key={spell.id} spell={spell} />
+                <DragSpell key={spell.id + spell.name} spell={spell} />
               ))}
             </div>
           </div>
